… js/web/test/data/ops/conv-transpose.jsonc