@xenova 2025-02-03
Not sure if this PR is related, but I am facing several correctness issues with ConvTranspose in https://github.com/microsoft/onnxruntime/issues/23403#issuecomment-2628883248.